requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
openai>=1.3.0
numpy>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
requests>=2.31.0
uuid>=1.30
watchdog==6.0.0
